requirements.txt
netCDF4==1.4.2
numpy==1.16.0
python-dateutil==2.7.5
six~=1.11.0
cumulus-process==0.8.0
